# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Le Kiosque Pédagogique API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StaffCredentials {
  /** @minLength 2 */
  username: string;
  password: string;
}

export interface School {
  id: number;
  name: string;
  /** @nullable */
  city?: string | null;
  active?: boolean;
}

export interface Discipline {
  id: number;
  name: string;
  active?: boolean;
}

export interface User {
  id: number;
  username: string;
  role: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  disciplineId?: number | null;
  active?: boolean;
  school?: School;
  discipline?: Discipline;
}

export interface AuthResult {
  token: string;
  user?: User;
}

export interface UserInput {
  /** @minLength 2 */
  username: string;
  password: string;
  role: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  disciplineId?: number | null;
}

export interface UserUpdate {
  /** @minLength 2 */
  username?: string;
  password?: string;
  role?: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  disciplineId?: number | null;
  active?: boolean;
}

export interface TicketInput {
  schoolId: number;
  disciplineId: number;
  description: string;
}

export interface TicketCredentials {
  ticketNumber: number;
  password: string;
}

export interface TicketCreated {
  ticketNumber: number;
  password: string;
  token: string;
}

export interface TicketSummary {
  id: number;
  status: string;
  /** @nullable */
  description?: string | null;
  schoolId?: number;
  disciplineId?: number;
  createdAt: string;
  updatedAt?: string;
  school?: School;
  discipline?: Discipline;
}

export interface TicketDetail {
  id: number;
  status: string;
  schoolId?: number;
  disciplineId?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  webexLink?: string | null;
  /** @nullable */
  webexScheduledAt?: string | null;
  /** @nullable */
  webexCreatedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
  school?: School;
  discipline?: Discipline;
  /**
     * Visible par le staff uniquement
     * @nullable
     */
  assignedN1Id?: number | null;
  /**
     * Visible par le staff uniquement
     * @nullable
     */
  assignedN2Id?: number | null;
}

export interface Colleague {
  id: number;
  username: string;
}

export interface ReassignInput {
  targetUserId: number;
}

export interface WebexInput {
  webexLink: string;
  /** Date et heure de la session visio */
  scheduledAt: string;
}

export interface Message {
  id: number;
  ticketId: number;
  /** @nullable */
  senderId?: number | null;
  senderType?: string;
  content: string;
  messageType: string;
  createdAt: string;
}

export interface MessageInput {
  ticketId: number;
  content: string;
}

export interface ResolutionStats {
  f2?: number;
  f3?: number;
  webex?: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DisciplineRanking {
  disciplineId: number;
  disciplineName: string;
  totalTickets: number;
  /** Percentage 0–100 of tickets remontés vers F3 */
  escalationRate: number;
  /** Percentage 0–100 of tickets résolus (F2, F3 ou Webex) */
  resolutionRate: number;
  /** @nullable */
  avgMinutes: number | null;
}

export interface RdDashboard {
  totalTickets: number;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  disciplineId?: number | null;
  /** @nullable */
  disciplineName?: string | null;
  /** @nullable */
  schoolName?: string | null;
  openTickets?: number;
  /** @nullable */
  avgPickupMinutes: number | null;
  resolutionByLevel: ResolutionStats;
  ticketsByStatus?: StatusCount[];
  /** Statistiques par discipline (vue direction, périmètre établissement) */
  disciplineRankings?: DisciplineRanking[];
}

export interface MonthlyCount {
  month: string;
  count: number;
}

export interface SchoolRanking {
  schoolId: number;
  schoolName: string;
  totalTickets: number;
  /** Percentage 0–100 of tickets remontés vers F3 */
  escalationRate: number;
  /** Percentage 0–100 of tickets résolus (F2, F3 ou Webex) */
  resolutionRate: number;
  /** @nullable */
  avgMinutes: number | null;
}

export interface PgDashboard {
  totalTickets: number;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  schoolName?: string | null;
  /** @nullable */
  disciplineId?: number | null;
  /** @nullable */
  disciplineName?: string | null;
  monthlyTrend: MonthlyCount[];
  schoolRankings: SchoolRanking[];
  disciplineRankings: DisciplineRanking[];
}

export interface SchoolInput {
  name: string;
  city?: string;
}

export interface DisciplineInput {
  name: string;
}

export type GetMyAssignedTicketsParams = {
/**
 * When true, returns resolved/closed tickets instead of in-progress ones
 */
resolved?: boolean;
};

export type GetDashboardRdParams = {
/**
 * Filtre par discipline (admin uniquement)
 */
disciplineId?: number;
/**
 * Filtre par établissement (admin uniquement)
 */
schoolId?: number;
};

export type ListRdTicketsParams = {
/**
 * Filtre par discipline (admin uniquement)
 */
disciplineId?: number;
/**
 * Filtre par établissement (admin uniquement)
 */
schoolId?: number;
status?: string;
};

export type GetDashboardPgParams = {
/**
 * Filtre par discipline (admin uniquement)
 */
disciplineId?: number;
/**
 * Filtre par établissement (admin uniquement)
 */
schoolId?: number;
};