# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Le Kiosque Pédagogique API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/staff-login:
    post:
      operationId: staffLogin
      tags: [auth]
      summary: Login for staff users (F2, F3, RD, PG, direction, admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # ── TICKETS ──────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a new staff user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Username already taken
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a staff user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
        "400":
          description: Cannot delete own account
  /admin/schools:
    post:
      operationId: createSchool
      tags: [admin]
      summary: Create a school (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "201":
          description: School created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /admin/disciplines:
    post:
      operationId: createDiscipline
      tags: [admin]
      summary: Create a discipline (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisciplineInput"
      responses:
        "201":
          description: Discipline created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discipline"
components:
  schemas:
...
    StaffCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 2
        password:
          type: string
    AuthResult:
      type: object
      required: [token]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        schoolId:
          type: ["integer", "null"]
        disciplineId:
          type: ["integer", "null"]
        active:
          type: boolean
        school:
          $ref: "#/components/schemas/School"
        discipline:
          $ref: "#/components/schemas/Discipline"
    UserInput:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 2
        password:
          type: string
        role:
          type: string
        schoolId:
          type: ["integer", "null"]
        disciplineId:
          type: ["integer", "null"]
    UserUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 2
        password:
          type: string
        role:
          type: string
        schoolId:
          type: ["integer", "null"]
        disciplineId:
          type: ["integer", "null"]
        active:
          type: boolean

    # ── TICKETS ──────────────────────────────────────────────────────────
...
    School:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: ["string", "null"]
        active:
          type: boolean
    SchoolInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        city:
          type: string
    Discipline:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        active:
          type: boolean
    DisciplineInput:
      type: object
      required: [name]
      properties:
        name:
          type: string