# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Le Kiosque Pédagogique API
 * OpenAPI spec version: 0.1.0
 */
import type { Discipline } from './discipline';
import type { School } from './school';

export interface User {
  id: number;
  username: string;
  role: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  disciplineId?: number | null;
  active?: boolean;
  school?: School;
  discipline?: Discipline;
}