# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Le Kiosque Pédagogique API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /** @minLength 2 */
  username: string;
  password: string;
  role: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  disciplineId?: number | null;
}